app/requirements.txt
Flask==2.1.2
setuptools==62.2.0
gunicorn==20.1.0
flask_sqlalchemy==2.5.1
PyJWT==1.4.0
Flask-JWT==0.3.2